refactor(tx-flow): move createTxFlow to tx-flow folder
