+++ apps/web/src/components/tx-flow-2/actions/ExecuteThroughRole.tsx
@@ -1,7 +1,7 @@
 import { SafeTxContext } from '@/components/tx-flow/SafeTxProvider'
 import { useCallback, useContext } from 'react'
 import { TxFlowContext } from '../TxFlowProvider'
-import type { SubmitCallback } from '../createTxFlow'
+import type { SubmitCallback } from '../../tx-flow/createTxFlow'
 import ExecuteThroughRoleForm from '@/components/tx/SignOrExecuteForm/ExecuteThroughRoleForm'
 import useIsCounterfactualSafe from '@/features/counterfactual/hooks/useIsCounterfactualSafe'
